+++ Changelog.md
@@ -1,6 +1,7 @@
 --------------------------Changelog for the next version of SecurityCraft--------------------------
 
 - New: In-game notification message that informs players that an error occurred while capturing a frame feed
+- New: Debug feature to help identify other mods preventing players from viewing a Security Camera
 - Fix: Several mod incompatibilities that caused Frame feeds to not show up correctly
 - Fix: Game performance does not improve when breaking frames with an active frame feed
 - Fix: Briefcase and Keycard Holder exploits